src/shared/domain/validators/__tests/unit/class-validator-fields.spec.ts
import { ClassValidatorFields } from '../../class-validator-fields';
import * as libClassValidator from 'class-validator';

class StubClassValidatorFields extends ClassValidatorFields<{
  field: string;
}> {}

describe('ClassValidatorFields unit test', () => {
  it('should initialize erros and validateDate as null', () => {
    const sut = new StubClassValidatorFields();

    expect(sut.errors).toBeNull();
    expect(sut.validatedData).toBeNull();
  });

  it('should validate with erros', () => {
    const spy = jest.spyOn(libClassValidator, 'validateSync');
    spy.mockReturnValue([
      { property: 'field', constraints: { isRequired: 'test error' } },
    ]);

    const sut = new StubClassValidatorFields();

    expect(sut.validate(null)).toBeFalsy();
    expect(spy).toBeCalled();

    expect(sut.errors).toStrictEqual({ field: ['test error'] });
  });
});
